+++ torch_geometric/nn/conv/wl_conv_continuous.py
@@ -3,8 +3,14 @@ from typing import Union
 from torch import Tensor
 
 from torch_geometric.nn.conv import MessagePassing
-from torch_geometric.typing import OptPairTensor, OptTensor, Size
-from torch_geometric.utils import scatter
+from torch_geometric.typing import (
+    Adj,
+    OptPairTensor,
+    OptTensor,
+    Size,
+    SparseTensor,
+)
+from torch_geometric.utils import scatter, spmm
 
 
 class WLConvContinuous(MessagePassing):
@@ -37,20 +43,31 @@ class WLConvContinuous(MessagePassing):
     def __init__(self, **kwargs):
         super().__init__(aggr='add', **kwargs)
 
-    def forward(self, x: Union[Tensor, OptPairTensor], edge_index: Tensor,
-                edge_weight: OptTensor = None, size: Size = None) -> Tensor:
+    def forward(
+        self,
+        x: Union[Tensor, OptPairTensor],
+        edge_index: Adj,
+        edge_weight: OptTensor = None,
+        size: Size = None,
+    ) -> Tensor:
 
         if isinstance(x, Tensor):
             x: OptPairTensor = (x, x)
 
-        if edge_weight is None:
-            edge_weight = x[0].new_ones(edge_index.size(1))
-
-        # propagate_type: (x: OptPairTensor, edge_weight: Tensor)
+        # propagate_type: (x: OptPairTensor, edge_weight: OptTensor)
         out = self.propagate(edge_index, x=x, edge_weight=edge_weight,
                              size=size)
 
-        deg = scatter(edge_weight, edge_index[1], 0, out.size(0), reduce='sum')
+        if isinstance(edge_index, SparseTensor):
+            assert edge_weight is None
+            dst_index, _, edge_weight = edge_index.coo()
+        else:
+            dst_index = edge_index[1]
+
+        if edge_weight is None:
+            edge_weight = x[0].new_ones(dst_index.numel())
+
+        deg = scatter(edge_weight, dst_index, 0, out.size(0), reduce='sum')
         deg_inv = 1. / deg
         deg_inv.masked_fill_(deg_inv == float('inf'), 0)
         out = deg_inv.view(-1, 1) * out
@@ -61,5 +78,12 @@ class WLConvContinuous(MessagePassing):
 
         return out
 
-    def message(self, x_j: Tensor, edge_weight: Tensor) -> Tensor:
-        return edge_weight.view(-1, 1) * x_j
+    def message(self, x_j: Tensor, edge_weight: OptTensor) -> Tensor:
+        return x_j if edge_weight is None else edge_weight.view(-1, 1) * x_j
+
+    def message_and_aggregate(
+        self,
+        adj_t: SparseTensor,
+        x: OptPairTensor,
+    ) -> Tensor:
+        return spmm(adj_t, x[0], reduce=self.aggr)